% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amphweb_data.R
\docType{data}
\name{amphweb}
\alias{amphweb}
\title{AmphibiaWeb Taxonomy}
\format{R dataset}
\usage{
amphweb
}
\description{
Data table of taxonomy downloaded from http://amphibiaweb.org/. The dataset contains a list of species and species names used by AmphibiaWeb as well as higher level taxonomy and more (IUCN red list status etc.). A version is stored internally for the purpose of running examples, but users should be aware that this may be outdated depending both on how recently the defrostR package as well as the AmphibiaWeb website has been updated.
}
\examples{
head(amphweb)
}
\keyword{datasets}
