% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asw_search.R
\name{asw_search}
\alias{asw_search}
\title{Search the Amphibian Species of the World database (Species search)}
\usage{
asw_search(query, asw_taxonomy = AmphiNom::asw_taxonomy)
}
\arguments{
\item{query}{search term. Binomial species name either separated by a space or "_"}

\item{asw_taxonomy}{version of taxonomy to use (output of get_taxonomy()). By default, internally stored version is used (which may be out of data)}
}
\value{
returns all available information in list form
}
\description{
This function searches species accounts on the ASW website and returns all available information listed. This currently works only at the species level, not higher taxonomic levels. The URL for a specific species account can be used as the query, or alternatively, the binomial species name. For the latter, unless specified otherwise, the function will look up the species' URL based on the information stored in AmphiNom's internal records, which may be outdated.
}
\examples{
asw_search(query="Pleurodeles waltl")
asw_search(query="https://amphibiansoftheworld.amnh.org/Amphibia/Caudata/Salamandridae/Pleurodelinae/Pleurodeles/Pleurodeles-waltl")
}
