% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asw_synonym_data.R
\docType{data}
\name{asw_synonyms}
\alias{asw_synonyms}
\title{Current version of Frost synonyms}
\format{R dataset}
\usage{
asw_synonyms
}
\description{
most current datasets
}
\examples{
head(asw_synonyms)
}
\keyword{datasets}
