% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asw_taxonomy_data.R
\docType{data}
\name{asw_taxonomy}
\alias{asw_taxonomy}
\title{Current version of Frost taxonomy}
\format{R dataset}
\usage{
asw_taxonomy
}
\description{
most current datasets
}
\examples{
head(asw_taxonomy)
}
\keyword{datasets}
