% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchASW.R
\name{searchASW}
\alias{searchASW}
\title{Search the Amphibian Species of the World database (Species search)}
\usage{
searchASW(query, asw_taxonomy = defrostR::asw_taxonomy)
}
\arguments{
\item{query}{search term. Binomial species name either separated by a space or "_"}

\item{asw_taxonomy}{version of taxonomy to use (output of getTaxonomy()). By default, internally stored version is used (which may be out of data)}
}
\value{
returns all available information in list form
}
\description{
This function searches species accounts on the ASW website (http://research.amnh.org/vz/herpetology/amphibia/index.php/) and returns all available information listed. The URL per species is taken from defrostR's internally stored records, unless specified otherwise.
}
\examples{
searchASW("Pleurodeles waltl")
}
