% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synonym_report.R
\name{synonym_report}
\alias{synonym_report}
\title{Get summary report of "synced" query}
\usage{
synonym_report(query_info, verbose = F)
}
\arguments{
\item{query_info}{output of the function asw_sync()}

\item{verbose}{logical arguments can be turned on to list species which result in conflict. Default=FALSE}
}
\value{
returns either a data frame or list of summary statitics
}
\description{
This function produces a report of an undertaken aswSync() query
}
\details{
One logical arguments can be turned on to summarize in terms of numbers of cases, or to provide species names of conflicting issues.
}
\examples{
#check out amphibiaweb taxonomy and extract only Bufonidae
head(amphweb$species)
amphweb_bufonidae<-amphweb[amphweb$family=="Bufonidae",]
#query bufonidae against ASW database and inspect
bufonidae_asw<-asw_sync(query=amphweb.bufonidae$species)
synonym_report(bufonidae_asw)
}
