% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amphweb_data.R
\docType{data}
\name{amphweb}
\alias{amphweb}
\title{Current version of Frost synonyms}
\format{R dataset}
\usage{
amphwebg
}
\description{
most current datasets
}
\examples{
head(amphweb)
}
\keyword{datasets}
