% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asw_synonym_data.R
\docType{data}
\name{asw_synonyms}
\alias{asw_synonyms}
\title{Amphibian Species of the World synonyms}
\format{R dataset}
\usage{
asw_synonyms
}
\description{
Data table generated using the getSynonyms() function. The dataset contains a full list of synonyms listed on the Amphibian Species of the World (ASW) website and the current accepted name of the species. A version is stored internally for the purpose of running examples, but users should be aware that this may be outdated depending both on how recently the defrostR package as well as the ASW website has been updated.
}
\examples{
head(asw_synonyms)
}
\keyword{datasets}
