% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asw_taxonomy_data.R
\docType{data}
\name{asw_taxonomy}
\alias{asw_taxonomy}
\title{Amphibian Species of the World Taxonomy}
\format{R dataset}
\usage{
asw_taxonomy
}
\description{
Data table generated using the getTaxonomy() function. The dataset contains a full list of species listed on the Amphibian Species of the World (ASW) website along with all higher level taxonomy and the URLs for each species. A version is stored internally for the purpose of running examples, but users should be aware that this may be outdated depending both on how recently the defrostR package as well as the ASW website has been updated.
}
\details{
dataset last updated on 12th July 2017
}
\examples{
head(asw_taxonomy)
str(asw_taxonomy)
asw_taxonomy<-load(asw_taxonomy)
}
\keyword{datasets}
